WIP rework to new brewblox-ctl

+++ brewblox_ctl_lib/lib_utils.py
@@ -3,13 +3,17 @@ Utility functions specific to lib
 """
 
 import re
-from subprocess import check_output
 
 import click
 import yaml
-from brewblox_ctl.utils import getenv, is_pi
 
-from brewblox_ctl_lib.const import HOST, HTTPS_PORT_KEY, SETENV
+from brewblox_ctl.utils import ctx_opts, getenv, is_pi
+from brewblox_ctl_lib.const import HOST, HTTPS_PORT_KEY, LOG_COMPOSE
+
+
+def is_dry():
+    ctx = click.get_current_context()
+    return ctx.ensure_object(dict).get('dry', False)
 
 
 def base_url():
@@ -40,8 +44,13 @@ def read_compose(fname='docker-compose.yml'):
 
 
 def write_compose(config, fname='docker-compose.yml'):  # pragma: no cover
-    with open(fname, 'w') as f:
-        yaml.safe_dump(config, f)
+    opts = ctx_opts()
+    if opts.dry_run or opts.verbose:
+        click.secho('{} {}'.format(LOG_COMPOSE, fname), fg='magenta', color=opts.color)
+
+    if not opts.dry_run:
+        with open(fname, 'w') as f:
+            yaml.safe_dump(config, f)
 
 
 def read_shared_compose(fname='docker-compose.shared.yml'):
@@ -59,14 +68,6 @@ def list_services(image=None, fname=None):
         k for k, v in config['services'].items()
         if image is None or v.get('image', '').startswith(image)
     ]
-
-
-def subcommand(cmd):  # pragma: no cover
-    return check_output(cmd, shell=True).decode()
-
-
-def setenv_cmd(key, value):
-    return '{} {} {}'.format(SETENV, key, value)
 
 
 def check_service_name(ctx, param, value):
